Added more test to cover charge method.
