+++ tests/unit/includes/gateway/abstract-omise-payment-offline-test.php
@@ -1,26 +1,13 @@
 <?php
 
-use PHPUnit\Framework\TestCase;
+require_once __DIR__ . '/bootstrap-test-setup.php';
 
-class Omise_Payment_Offline_Test extends TestCase
+class Omise_Payment_Offline_Test extends Bootstrap_Test_Setup
 {
     public function setUp(): void
     {
+        parent::setUp();
         require_once __DIR__ . '/../../../../includes/gateway/abstract-omise-payment-offline.php';
-    }
-
-    public function getOrderMock($expectedAmount, $expectedCurrency)
-    {
-        // Create a mock of the $order object
-        $orderMock = Mockery::mock('WC_Order');
-
-        // Define expectations for the mock
-        $orderMock->shouldReceive('get_currency')
-            ->andReturn($expectedCurrency);
-        $orderMock->shouldReceive('get_total')
-            ->andReturn($expectedAmount);  // in units
-        $orderMock->shouldReceive('add_meta_data');
-        return $orderMock;
     }
 
     public function charge()
@@ -31,6 +18,5 @@ class Omise_Payment_Offline_Test extends TestCase
         $result = $mock->charge('order_123', $orderMock);
 
         var_dump(print_r($result, true));
-
     }
 }
