+++ tests/unit/includes/gateway/class-omise-payment-ocbc-digital-test.php
@@ -88,33 +88,6 @@ class Omise_Payment_OCBC_Digital_Test extends Omise_Offsite_Test
      */
     public function testCharge()
     {
-        $expectedCurrency = 'SGD';
-        $expectedAmount = 1000000; // in subunits
-        $expectedAmount = 1000000; // in subunits
-        $expectedRequest = [
-            "object" => "charge",
-            "id" => "chrg_test_no1t4tnemucod0e51mo",
-            "location" => "/charges/chrg_test_no1t4tnemucod0e51mo",
-            "amount" => $expectedAmount,
-            "currency" => $expectedCurrency
-        ];
-
-        // Create a mock for OmiseCharge
-        $chargeMock = Mockery::mock('overload:OmiseCharge');
-        $chargeMock->shouldReceive('create')->once()->andReturn($expectedRequest);
-
-        // Create a mock of the $order object
-        $orderMock = Mockery::mock('WC_Order');
-
-        // Define expectations for the mock
-        $orderMock->shouldReceive('get_currency')
-            ->andReturn($expectedCurrency);
-        $orderMock->shouldReceive('get_total')
-            ->andReturn($expectedAmount/100);  // in units
-
-        $order_id = "123";
-        $result = $this->obj->charge($order_id, $orderMock);
-        $this->assertEquals($expectedAmount, $result['amount']);
-        $this->assertEquals($expectedCurrency, $result['currency']);
+        $this->getChargeTest($this->obj);
     }
 }
